Secure-File-Transfer/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
bcrypt
sqlalchemy
cryptography
PyJWT
python-jose[cryptography]
